# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Distribuidora API",
    "version": "dev"
  },
  "paths": {
...
    "/api/v1/configuracion/alicuotas": {
      "get": {
        "tags": [
          "configuracion"
        ],
        "summary": "Listar Alicuotas",
        "operationId": "listar_alicuotas_api_v1_configuracion_alicuotas_get",
        "parameters": [
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaAlicuotas"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configuracion/medios-pago": {
      "get": {
        "tags": [
          "configuracion"
        ],
        "summary": "Listar Medios Pago",
        "operationId": "listar_medios_pago_api_v1_configuracion_medios_pago_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaMediosPago"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/configuracion/motivos": {
      "get": {
        "tags": [
          "configuracion"
        ],
        "summary": "Listar Motivos",
        "operationId": "listar_motivos_api_v1_configuracion_motivos_get",
        "parameters": [
          {
            "name": "ambito",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ambito"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListaMotivos"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlicuotaResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "valor": {
            "type": "string",
            "title": "Valor"
          },
          "activo": {
            "type": "boolean",
            "title": "Activo"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "valor",
          "activo"
        ],
        "title": "AlicuotaResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ListaMediosPago": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MedioPagoResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ListaMediosPago"
      },
      "ListaMotivos": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MotivoResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ListaMotivos"
      },
...
      "MedioPagoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "requiere_referencia": {
            "type": "boolean",
            "title": "Requiere Referencia"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "requiere_referencia"
        ],
        "title": "MedioPagoResponse",
        "description": "Un medio de pago activo con su indicador de referencia obligatoria."
      },
      "MotivoResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre"
        ],
        "title": "MotivoResponse"
      },
...
      "PaginaAlicuotas": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlicuotaResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "cursor_siguiente": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cursor Siguiente"
          }
        },
        "type": "object",
        "required": [
          "items",
          "cursor_siguiente"
        ],
        "title": "PaginaAlicuotas"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}